adding new shared util file for longview, removing entity icon from IconSection
