+++ packages/manager/src/features/Longview/LongviewLanding/Gauges/Storage.tsx
@@ -1,5 +1,4 @@
 import { APIError } from 'linode-js-sdk/lib/types';
-import { pathOr } from 'ramda';
 import * as React from 'react';
 import Typography from 'src/components/core/Typography';
 import GaugePercent from 'src/components/GaugePercent';
@@ -7,7 +6,7 @@ import withClientStats, {
   Props as LVDataProps
 } from 'src/containers/longview.stats.container';
 import { readableBytes } from 'src/utilities/unitConversions';
-import { Disk } from '../../request.types';
+import { sumStorage } from '../../shared/utilities';
 import { baseGaugeProps, BaseProps as Props } from './common';
 
 const StorageGauge: React.FC<Props & LVDataProps> = props => {
@@ -52,23 +51,6 @@ const StorageGauge: React.FC<Props & LVDataProps> = props => {
 };
 
 export default withClientStats<Props>(props => props.clientID)(StorageGauge);
-
-// UTILITIES
-interface Storage {
-  free: number;
-  total: number;
-}
-
-export const sumStorage = (DiskData: Record<string, Disk> = {}): Storage => {
-  let free = 0;
-  let total = 0;
-  Object.keys(DiskData).forEach(key => {
-    const disk = DiskData[key];
-    free += pathOr(0, ['fs', 'free', 0, 'y'], disk);
-    total += pathOr(0, ['fs', 'total', 0, 'y'], disk);
-  });
-  return { free, total };
-};
 
 export const innerText = (
   value: string,
