+++ packages/manager/src/features/Longview/shared/utilities.ts
@@ -1,0 +1,29 @@
+import { pathOr } from 'ramda';
+import { pluralize } from 'src/utilities/pluralize';
+import { Disk, LongviewPackage } from '../request.types';
+
+interface Storage {
+  free: number;
+  total: number;
+}
+
+export const getPackageNoticeText = (packages: LongviewPackage[]) => {
+  if (!packages) {
+    return 'Package information not available';
+  }
+  if (packages.length === 0) {
+    return 'All packages up to date';
+  }
+  return `${pluralize('package', 'packages', packages.length)} have updates`;
+};
+
+export const sumStorage = (DiskData: Record<string, Disk> = {}): Storage => {
+  let free = 0;
+  let total = 0;
+  Object.keys(DiskData).forEach(key => {
+    const disk = DiskData[key];
+    free += pathOr(0, ['fs', 'free', 0, 'y'], disk);
+    total += pathOr(0, ['fs', 'total', 0, 'y'], disk);
+  });
+  return { free, total };
+};
